1.修改动画弹出窗口caption失效问题和取消第一个TAB控件作焦点 2.增加RichEdit控件，剪切和复制delete时异常暂时未坚决

+++ Duilib/Control/DuiFadeButton.cpp
@@ -76,7 +76,7 @@ namespace DuiLib
         }
         if(event.Type == UIEVENT_TIMER)
         {
-            OnTimer(event.wParam);
+            OnTimer((int)event.wParam);
         }
         CDuiButton::DoEvent(event);
     }
